backend/pyproject.toml
[project]
name = "winaity-extractor-backend"
version = "0.1.0"
description = "Winaity Extractor - B2B Lead Extraction API"
requires-python = ">=3.12"
dependencies = [
    # Framework
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.34,<1.0",
    "pydantic>=2.10,<3.0",
    "pydantic-settings>=2.7,<3.0",
    # Database
    "sqlalchemy[asyncio]>=2.0,<3.0",
    "asyncpg>=0.30,<1.0",
    "alembic>=1.14,<2.0",
    # Redis
    "redis[hiredis]>=5.2,<6.0",
    # Celery
    "celery[redis]>=5.4,<6.0",
    # Auth
    "python-jose[cryptography]>=3.3,<4.0",
    "passlib[bcrypt]>=1.7,<2.0",
    "python-multipart>=0.0.18,<1.0",
    # Stripe
    "stripe>=11.0,<12.0",
    # Validation
    "email-validator>=2.2,<3.0",
    "phonenumbers>=8.13,<9.0",
    "python-slugify>=8.0,<9.0",
    # Export
    "openpyxl>=3.1,<4.0",
    # HTTP
    "httpx>=0.28,<1.0",
    # AWS
    "boto3>=1.35,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
    "pytest-asyncio>=0.24,<1.0",
    "httpx>=0.28,<1.0",
    "factory-boy>=3.3,<4.0",
    "ruff>=0.8,<1.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
